refactor: split deploy coordinator into mutex and replica-routing actors
Gateway watch-loop traffic (per-replica long-poll) shared an actor and
event loop with the low-frequency cross-operator deploy mutex. Splits
the routing registry (register/unregister/set_expected/get_routing/
wait_for_change + StateStore persistence) into a new ReplicaCoordinator
actor, so gateway replica count no longer contends with deploy locking.
No behavior change; DeployCoordinator keeps the mutex/liveness concern.
