+++ tests/test_api.py
@@ -216,7 +216,7 @@ class TestWatchReconcile:
             "generation": 3,
         }
         with (
-            patch("modelship.infer.deploy_coordinator.get_or_create_coordinator", return_value=MagicMock()),
+            patch("modelship.infer.replica_coordinator.get_or_create_replica_coordinator", return_value=MagicMock()),
             patch("modelship.openai.api.ray.get", return_value=snapshot),
             patch("modelship.openai.api.serve.get_app_handle", return_value=MagicMock()),
         ):
@@ -228,7 +228,7 @@ class TestWatchReconcile:
 
     def test_sync_tolerates_unavailable_coordinator(self, api):
         api._watch_task = None
-        with patch("modelship.infer.deploy_coordinator.get_or_create_coordinator", side_effect=RuntimeError):
+        with patch("modelship.infer.replica_coordinator.get_or_create_replica_coordinator", side_effect=RuntimeError):
             assert api._sync_routing_blocking() is False
         assert api.models == {}
 
@@ -239,7 +239,7 @@ class TestWatchReconcile:
         api._watch_task = None
         snapshot = {"models": {"qwen-aaaaaaaaaa": "qwen"}, "expected": ["qwen"], "generation": 2}
         with (
-            patch("modelship.infer.deploy_coordinator.get_or_create_coordinator", return_value=MagicMock()),
+            patch("modelship.infer.replica_coordinator.get_or_create_replica_coordinator", return_value=MagicMock()),
             patch("modelship.openai.api.ray.get", return_value=snapshot),
             patch("modelship.openai.api.serve.get_app_handle", side_effect=RuntimeError("controller lag")),
         ):
@@ -252,18 +252,20 @@ class TestWatchReconcile:
         # cleared so the next _coord() re-resolves instead of retrying a corpse.
         stale = MagicMock()
         stale.get_routing.remote.side_effect = RuntimeError("actor dead")
-        api._coordinator = stale
+        api._replica_coord = stale
         with patch("modelship.openai.api.ray.get", side_effect=RuntimeError("actor dead")):
             assert api._sync_routing_blocking() is False
-        assert api._coordinator is None
+        assert api._replica_coord is None
 
     @pytest.mark.asyncio
     async def test_coord_async_resolves_off_thread_and_caches(self, api):
         # The watch loop resolves the coordinator via asyncio.to_thread (so the sync
         # ray.get_actor never blocks the event loop) and caches the handle.
-        api._coordinator = None
+        api._replica_coord = None
         sentinel = MagicMock()
-        with patch("modelship.infer.deploy_coordinator.get_or_create_coordinator", return_value=sentinel) as goc:
+        with patch(
+            "modelship.infer.replica_coordinator.get_or_create_replica_coordinator", return_value=sentinel
+        ) as goc:
             assert await api._coord_async() is sentinel
             assert await api._coord_async() is sentinel
         goc.assert_called_once()  # second call served from cache, no re-resolve
@@ -274,10 +276,10 @@ class TestWatchReconcile:
         _apply(api, {"qwen-aaaaaaaaaa": "qwen"}, gen=4)
         empty = {"models": {}, "expected": [], "generation": 0}
         with (
-            patch("modelship.infer.deploy_coordinator.get_or_create_coordinator", return_value=MagicMock()),
+            patch("modelship.infer.replica_coordinator.get_or_create_replica_coordinator", return_value=MagicMock()),
             patch("modelship.openai.api.ray.get", return_value=empty),
         ):
-            api._coordinator = None
+            api._replica_coord = None
             assert api._sync_routing_blocking() is True
         assert "qwen" in api.models
